back-end/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.10.1
